we have the first index insert working now

+++ input/javasrc/com/alvazan/orm/layer3/spi/db/cassandra/BigIntegerComposite.java
@@ -1,13 +1,13 @@
 package com.alvazan.orm.layer3.spi.db.cassandra;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import com.netflix.astyanax.annotations.Component;
 
 public class BigIntegerComposite {
 
 	@Component(ordinal=0)
-	BigDecimal value;
+	BigInteger value;
 	@Component(ordinal=1)
 	byte[] pk;
